+++ adapter/spark/src/main/java/com/asml/apa/wta/spark/Main.java
@@ -1,7 +1,6 @@
 package com.asml.apa.wta.spark;
 
 public class Main {
-  public static void main(String[] args) {
-    // System.out.println("Hello World!");
-  }
+
+  public static void main(String[] args) {}
 }